reorganize project to split table/ and semantic_table/
