+++ corvid/table_aggregation/schema_matcher.py
@@ -3,8 +3,7 @@ from typing import List
 import numpy as np
 from scipy.optimize import linear_sum_assignment
 
-from corvid.semantic_table.table import Table, Cell
-from corvid.semantic_table.semantic_table import SemanticTable
+from corvid.table.table import Table, Cell
 from corvid.table_aggregation.pairwise_mapping import PairwiseMapping
